Separate classes + more unity-like UX on yarn assets
+ Put runtime classes and editor classes into separate files
+ Editor class of a yarn asset inherits from ScriptedImporterEditor instead of Editor which means changes on the Asset will be applied by clicking either the "Apply" or "Revert" button making this more unity-like UX wise
+ Move Culture struct outside of YarnImporterEditor so it can be used in other classes too
