+++ Unity/Assets/YarnSpinner/Editor/YarnImporterEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 86e7a1afcf36f2847b546e40c89ce697
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
