+++ Unity/Assets/YarnSpinner/Editor/YarnProgramEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 76b0f75fe69847346ad55b49ea4197ea
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
