+++ Unity/Assets/YarnSpinner/Scripts/Culture.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+[Serializable]
+public struct Culture {
+    public string Name;
+    public string DisplayName;
+}
