+++ Unity/Assets/YarnSpinner/Scripts/Culture.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b333bea34d842c14592b64107ba0ada8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
