+++ Unity/Assets/YarnSpinner/Scripts/YarnProgram.cs
@@ -3,11 +3,6 @@ using System.Collections.Generic;
 using UnityEngine;
 using Yarn;
 
-#if UNITY_EDITOR
-using UnityEditor;
-using System.Linq;
-#endif
-
 /// Stores compiled Yarn programs in a form that Unity can serialise.
 public class YarnProgram : ScriptableObject
 {
@@ -24,44 +19,3 @@ public class YarnProgram : ScriptableObject
         return Program.Parser.ParseFrom(compiledProgram);                
     } 
 }
-
-#if UNITY_EDITOR
-/// A custom editor that lists the nodes in the program.
-[CustomEditor(typeof(YarnProgram))]
-public class YarnProgramEditor : Editor {
-
-    private List<string> nodeNames;
-
-    void OnEnable() {
-        try {
-            var program = (serializedObject.targetObject as YarnProgram).GetProgram();
-
-            nodeNames = program.Nodes.Keys.ToList();
-        } catch (YarnException) {
-            nodeNames = null;
-        }
-        
-    }
-
-    public override void OnInspectorGUI() {
-        base.OnInspectorGUI();
-
-        if (nodeNames == null) {
-            EditorGUILayout.HelpBox("Error reading Yarn program.", MessageType.Error);
-            return;
-        }
-        
-        EditorGUILayout.LabelField($"{nodeNames.Count} nodes:", EditorStyles.boldLabel);
-        EditorGUI.indentLevel += 1;
-
-        foreach (var nodeName in nodeNames) {
-            EditorGUILayout.LabelField(nodeName);
-        }
-
-        EditorGUI.indentLevel -= 1;
-        
-    }
-}
-#endif
-
-
